requirements.txt
ctransformers
sentence-transformers
langchain
langchain-community
pinecone-client
flask
pypdf
python-dotenv

-e .